requirements.txt
pandas
numpy
matplotlib
seaborn
requests
pyyaml
scikit-learn
jupyterlab
ipykernel
python-dotenv
